+++ tests/assets/bundler/main.tsx
@@ -1,9 +1,9 @@
 /// <reference lib="dom" />
+import Title from "mod/title.tsx";
+import { salute } from "mod/utils.ts";
 import React from "https://deno.land/x/react_deno@17.0.2/react.ts";
 import ReactDOM from "https://deno.land/x/react_deno@17.0.2/dom.ts";
-import Title from "./modules/title.tsx";
-import { salute } from "./utils.ts";
 
-const content = salute("Óscar");
+const content = salute("Oscar");
 
 ReactDOM.render(<Title salute={content} />, document.getElementById("app"));